requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
opencv-python
Pillow
onnx
onnxsim
onnxruntime
tqdm
tensorboardX